### Motivation / Background

This Pull Request has been created because https://github.com/rails/rails/pull/52664 rightfully reverted https://github.com/rails/rails/pull/52512 and https://github.com/rails/rails/pull/52605 because it didn't have proper backward compatibility with hash-based routes. This PR is another attempt at keyword routes that supports hashes too.

### Detail

This Pull Request changes the routing mapper to use keywords, and forwards hash keys to keyword values.

### Additional information

Here's trivial benchmark indicating the performance cost compared to main:

Benchmark:
```rb
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", path: "."
  gem "benchmark-ips"
  gem "stringio"
end

require "benchmark/ips"
require "action_controller/railtie"

class TestApp < Rails::Application
  config.root = __dir__
  config.hosts << "example.org"
  config.secret_key_base = "secret_key_base"

  config.logger = Logger.new($stdout)
  Rails.logger  = config.logger
end

class TestController < ActionController::Base
  include Rails.application.routes.url_helpers

  def index
    render plain: "Home"
  end
end

Benchmark.ips do |x|
  x.report("draw") do
    Rails.application.routes.draw do
      resources :posts
      resource :user
      root to: "home#index"
      post "login", to: "sessions#login"
      delete "logout", to: "sessions#logout"
    end
    Rails.application.instance_variable_set(:@routes, nil)
  end
end
```

Main:
```
Warming up --------------------------------------
                draw   146.000 i/100ms
Calculating -------------------------------------
                draw      1.443k (± 3.1%) i/s  (692.77 μs/i) -      7.300k in   5.062232s
```

This branch:
```
Warming up --------------------------------------
                draw   143.000 i/100ms
Calculating -------------------------------------
                draw      1.440k (± 2.8%) i/s  (694.68 μs/i) -      7.293k in   5.070202s
```
Same-ish performance, I think.

This branch without hash support (post deprecation cycle):
```
Warming up --------------------------------------
                draw   148.000 i/100ms
Calculating -------------------------------------
                draw      1.459k (± 1.9%) i/s  (685.38 μs/i) -      7.400k in   5.073486s
```
It seems very marginally faster.

Though, this change's primary motivation is not speed, it is maintainability. The flow of each mapper method is much easier to follow with keywords.

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.