@claudiob 2025-04-14
Thanks @gmcgibbon for the contribution!

I have a question regarding how to get rid of the deprecation warning that shows after this commit.

My route files looks likes this (simplified):

```ruby
options = { param: :identifier, constraints: { identifier: /[abcdef]{6}/ } }

resources :locations, options.merge(only: [:edit, :update])
resources :users do
  resources :locations, options.merge(only: :show)
end
```

In other words, whenever I access a `location` resource, I don't want to use its primary key (ID) but a custom identifier stored in a different column, so my URLs look like this:

- `GET /locations/aecdec/edit` > edits a location
- `PATCH /locations/aecdec` > updates a location
- `GET /users/4/locations/aecdec` > shows the location of a user

---

After this commit I am getting a deprecation warning: `resources received a hash argument param. Please use a keyword instead`.

I'm looking at the new `resources` method but it's not fully clear to me how to update my own code to avoid the deprecation warning. Do you have any suggestions? Thank you! 🙇 

https://github.com/rails/rails/blob/3b4255e1800d0b53c96db5aeb88aaa66163f74d5/actionpack/lib/action_dispatch/routing/mapper.rb#L1667-L1671