@skipkayhil 2025-04-14
You should splat your options hashes so they are passed as keyword args:

```ruby
options = { param: :identifier, constraints: { identifier: /[abcdef]{6}/ } }

resources :locations, **options.merge(only: [:edit, :update])
resources :users do
  resources :locations, **options.merge(only: :show)
end
```